requirements.txt
pandas
numpy
seaborn
pywavelets==1.4.1
scikit-learn
scip
logger
matplotlib
#-e .
